refactor: 더보기 모달시 삭제 기능 수정
