--- src/utils/formatDate.ts
@@ -1,15 +1,0 @@
-export const formatDate = (dateString?: string): string => {
-  if (typeof dateString !== 'string') return '';
-
-  const date = new Date(dateString);
-
-  const year = date.getFullYear();
-  const month = `${date.getMonth() + 1}`.padStart(2, '0');
-  const day = `${date.getDate()}`.padStart(2, '0');
-
-  const hours = `${date.getHours()}`.padStart(2, '0');
-  const minutes = `${date.getMinutes()}`.padStart(2, '0');
-  const seconds = `${date.getSeconds()}`.padStart(2, '0');
-
-  return `${year}-${month}-${day} ${hours}:${minutes}:${seconds}`;
-};
